+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - Re-install runAccountsSelection() client function, thus bumping minor candidate version number
     - Fix typo in 'pwix.AccountsCore.m.createAccount' method
     - Fix AccountsCore.s.byQuery() calls with missing acInstance
+    - Define new 'preferredLabelWarnsOnce' configuration parameter, and honor it
 
 ### 2.2.2